refactor: DomainFixture를 domain패키지 하위로 이동시킨다
