+++ src/test/java/subway/domain/DomainFixture.java
@@ -1,6 +1,4 @@
-package subway;
-
-import subway.domain.Station;
+package subway.domain;
 
 public class DomainFixture {
