+++ src/test/java/subway/domain/LineTest.java
@@ -9,7 +9,6 @@ import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
-import subway.DomainFixture;
 
 @DisplayName("Line 클래스")
 class LineTest {
