+++ src/test/java/subway/domain/SectionTest.java
@@ -8,7 +8,6 @@ import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
-import subway.DomainFixture;
 
 @DisplayName("Section 클래스")
 class SectionTest {
